requirements.txt
asgiref==3.2.3
asn1crypto==0.24.0
attrs==17.4.0
autobahn==20.3.1
Automat==0.6.0
Babel==2.8.0
blinker==1.4
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.0
channels==2.4.0
chardet==3.0.4
click==6.7
cloud-init==19.4
colorama==0.3.7
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
daphne==2.4.1
defusedxml==0.6.0
distro-info===0.18ubuntu0.18.04.1
Django==3.0.4
django-bootstrap-modal-forms==1.5.0
django-channels==0.7.0
django-classy-tags==1.0.0
django-crispy-forms==1.8.1
django-debug-toolbar==2.1
django-json-widget==1.0.0
django-phonenumber-field==4.0.0
django-places==0.1
django-pwa==1.0.8
django-rest-framework==0.1.0
django-sekizai==1.1.0
django-sslserver==0.22
django-storages==1.9.1
django-user-agents==0.4.0
django-webpush==0.3.1
django-widget-tweaks==1.4.5
djangorestframework==3.11.0
ec2-hibinit-agent==1.0.0
future==0.18.2
geographiclib==1.50
geopy==1.20.0
google-api-core==1.16.0
google-auth==1.11.2
google-cloud-core==1.3.0
google-cloud-storage==1.26.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
hibagent==1.0.1
http-ece==1.1.0
httplib2==0.9.2
hyperlink==17.3.1
idna==2.8
incremental==16.10.1
Jinja2==2.10
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
MarkupSafe==1.0
mysqlclient==1.4.6
netifaces==0.10.4
oauthlib==3.1.0
PAM==0.4.2
phonenumbers==8.11.2
Pillow==7.0.0
places==0.3.4
protobuf==3.11.3
psycopg2-binary==2.8.4
py-vapid==1.7.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.20
pycrypto==2.6.1
pygobject==3.26.1
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==17.5.0
pyserial==3.4
python-apt==1.6.5+ubuntu0.2
python-debian==0.1.32
python-decouple==3.3
python-memcached==1.59
python3-openid==3.1.0
pytz==2019.3
pywebpush==1.9.4
pyxdg==0.25
PyYAML==3.12
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
rsa==4.0
SecretStorage==2.3.1
service-identity==18.1.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
ssh-import-id==5.7
systemd-python==234
Twisted==19.10.0
txaio==20.1.1
ua-parser==0.10.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.7
user-agents==2.1
zope.interface==4.7.2
